Use generic device for tile/switch/multizone ops, not "bulb".
It would probably be "better" to create an individual device class for each one, but that requires an additional call to the device to figure out what it is, which seems to be a bit out of scope for the library currently.

So, instead, we'll just use the generic "device".